+++ lang/en/chats.php
@@ -9,6 +9,9 @@ return [
         'heading' => 'Chats',
         'no_conversations_yet' => 'No conversations yet',
         'you' => 'You',
+        'attachment' => 'Attachment',
+        'now' => 'Now',
+        'load_more' => 'Load more',
 
     ],