+++ lang/en/validation.php
@@ -1,0 +1,15 @@
+<?php
+
+return [
+    // Laravel syntax
+    'file' => 'The :attribute field must be a file.',
+    'required' => 'The :attribute field is required.',
+    'max' => [
+        'file' => 'The :attribute must not be greater than :max kilobytes.',
+        'string' => 'The :attribute field must not be greater than :max characters.',
+    ],
+    'mimes' => 'The :attribute field must be a file of type: :values.',
+    'image' => 'The :attribute field must be an image.',
+    // Custom
+
+];