+++ resources/views/livewire/chats/partials/load-more-button.blade.php
@@ -2,7 +2,7 @@
     <button wire:loading.remove wire:target="loadMore" wire:loading.attr="disabled"
         dusk="loadMoreButton" @click="$wire.loadMore()"
         class="  text-sm dark:text-white disabled:hover:cursor-not-allowed hover:text-gray-700 transition-colors dark:hover:text-gray-500 dark:gray-200">
-        Load more
+        @lang('wirechat::chats.labels.load_more')
     </button>
 
     <div wire:loading wire:target="loadMore">