+++ resources/views/livewire/chats/partials/message-body.blade.php
@@ -3,7 +3,7 @@
     {{-- Only show if AUTH is onwer of message --}}
     @if ($belongsToAuth)
         <span class="font-bold text-xs dark:text-white/90 dark:font-normal">
-            You:
+            @lang('wirechat::chats.labels.you'):
         </span>
     @elseif(!$belongsToAuth && $group !== null)
         <span class="font-bold text-xs dark:text-white/80 dark:font-normal">
@@ -20,12 +20,12 @@
         'font-normal text-gray-600' =>
             $isReadByAuth && $lastMessage?->ownedBy($this->auth),
     ])>
-        {{ $lastMessage->body != '' ? $lastMessage->body : ($lastMessage->isAttachment() ? '📎 Attachment' : '') }}
+        {{ $lastMessage->body != '' ? $lastMessage->body : ($lastMessage->isAttachment() ? '📎 '.__('wirechat::chats.labels.attachment') : '') }}
     </p>
 
     <span class="font-medium px-1 text-xs shrink-0 text-gray-800 dark:text-gray-50">
         @if ($lastMessage->created_at->diffInMinutes(now()) < 1)
-            now
+          @lang('wirechat::chats.labels.now')
         @else
             {{ $lastMessage->created_at->shortAbsoluteDiffForHumans() }}
         @endif